layout_focus_group {graphlayouts}	R Documentation

radial focus group layout

Description

arrange nodes in concentric circles around a focal node according to their distance from the focus and keep predefined groups in the same angle range.

Usage

layout_with_focus_group(
  g,
  v,
  group,
  shrink = 10,
  weights = NA,
  iter = 500,
  tol = 1e-04
)

layout_igraph_focus_group(
  g,
  v,
  group,
  shrink = 10,
  weights = NA,
  iter = 500,
  tol = 1e-04,
  circular
)

Arguments

g
igraph object

v
id of focal node to be placed in the center

group
vector indicating grouping of nodes

shrink
shrink the reserved angle range for a group to increase the gaps between groups

weights
possibly a numeric vector with edge weights. If this is NULL and the graph has a weight edge attribute, then the attribute is used. If this is NA then no weights are used (even if the graph has a weight attribute). By default, weights are ignored. See details for more.

iter
number of iterations during stress optimization

tol
stopping criterion for stress optimization

circular
not used

Details

Be careful when using weights. In most cases, the inverse of the edge weights should be used to ensure that the endpoints of an edges with higher weights are closer together (weights=1/E(g)$weight).

Value

matrix of xy coordinates

See Also

layout_focus The layout_igraph_* function should not be used directly. It is only used as an argument for plotting with 'igraph'.

Examples

library(igraph)
g <- sample_islands(4, 5, 0.8, 2)
grp <- as.character(rep(1:4, each = 5))
layout_with_focus_group(g, v = 1, group = grp, shrink = 10)

[Package graphlayouts version 1.2.3 Index]