Calculates the orbit-aware quad census on an edge and node level, see vignette('oaqc').

Calculates the orbit-aware quad census on an edge and node level, see vignette('oaqc').

Usage

oaqc(graph, non_ind_freq = F, file = "")

Arguments

graph

A matrix, data.frame or graph object.

non_ind_freq

A flag indicating whether non-induced frequencies have to be returned or not.

file

Name (and location) of the file to be written.

Value

orbit-aware quad census on a node and edge level. Consult vignette('oaqc') to see the correspondence between orbit and quad.

Examples

k4 <- data.frame(
    source = c(0, 0, 0, 1, 1, 2),
    target = c(1, 2, 3, 2, 3, 3)
)

k4orbits <- oaqc(k4, non_ind_freq = TRUE)
print(k4orbits)
#> $n_orbits_non_ind
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    3    3    3    3    6    3    3    1     6     6     1     3     3
#> [2,]    1    3    3    3    3    6    3    3    1     6     6     1     3     3
#> [3,]    1    3    3    3    3    6    3    3    1     6     6     1     3     3
#> [4,]    1    3    3    3    3    6    3    3    1     6     6     1     3     3
#>      [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]     6     3     3     3     3     1
#> [2,]     6     3     3     3     3     1
#> [3,]     6     3     3     3     3     1
#> [4,]     6     3     3     3     3     1
#> 
#> $e_orbits_non_ind
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> [2,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> [3,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> [4,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> [5,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> [6,]    1    1    4    2    4    2    2    2    4     2     2     4     1     1
#> 
#> $n_orbits_ind
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> [2,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#> [4,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]     0     0     0     0     0     1
#> [2,]     0     0     0     0     0     1
#> [3,]     0     0     0     0     0     1
#> [4,]     0     0     0     0     0     1
#> 
#> $e_orbits_ind
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#> [2,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#> [3,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#> [4,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#> [5,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#> [6,]    0    0    0    0    0    0    0    0    0     0     0     0     0     1
#>