graphlayouts: layout algorithms for network visualizations

The package implements several new layout algorithms to visualize networks. Most are based on the concept of stress majorization. Some more specific algorithms allow to emphasize hidden group structures in networks or focus on specific nodes. The package is best used in conjunction with ggraph.

Some features of the package are:

layout_with_stress() is a state of the art deterministic layout algorithms.

layout_as_backbone() uncovers hidden group structures (if they exist) by emphasizing strongly embedded edges.

layout_with_focus() and layout_with_centrality() produce concentric layouts with a focal or most central nodes in the center.

layout_with_eigen() implements some layout algorithms on the basis of eigenvectors

layout_with_sparse_stress() sparse stress for large graphs

layout_with_pmds() pivot MDS for large graphs.

layout_as_dynamic() for longitudinal network data

A detailed tutorial can be found at https://schochastics.github.io/netVizR/

See also

Author

Maintainer: David Schoch david@schochastics.net (ORCID)

Other contributors:

Mark Ortmann (author of R package oaqc) [contributor]

Felix Schoenenberger (author of R package oaqc) [contributor]