dynamic graph layout

Create layouts for longitudinal networks.

Usage

layout_as_dynamic(gList, weights = NA, alpha = 0.5, iter = 500, tol = 1e-04)

Arguments

gList

list of igraph objects. Each network must contain the same set of nodes.

weights

possibly a numeric vector with edge weights. If this is NULL and the graph has a weight edge attribute, then the attribute is used. If this is NA then no weights are used (even if the graph has a weight attribute). By default, weights are ignored. See details for more.

alpha

weighting of reference layout. See details.

iter

number of iterations during stress optimization

tol

stopping criterion for stress optimization

Value

list of coordinates for each graph

Details

The reference layout is calculated based on the union of all graphs. The parameter alpha controls the influence of the reference layout. For alpha=1, only the reference layout is used and all graphs have the same layout. For alpha=0, the stress layout of each individual graph is used. Values in-between interpolate between the two layouts.

Be careful when using weights. In most cases, the inverse of the edge weights should be used to ensure that the endpoints of an edges with higher weights are closer together (weights=1/E(g)$weight).

References

Brandes, U. and Indlekofer, N. and Mader, M. (2012). Visualization methods for longitudinal social networks and stochastic actor-oriented modeling. Social Networks 34 (3) 291-308

Examples

library(igraph)
g1 <- sample_gnp(20, 0.2)
g2 <- sample_gnp(20, 0.2)
g3 <- sample_gnp(20, 0.2)

xy <- layout_as_dynamic(list(g1, g2, g3))

# layout for first network
xy[[1]]
#>              [,1]        [,2]
#>  [1,]  0.55919013  1.37581901
#>  [2,]  1.23018164  1.85635297
#>  [3,] -1.09703533 -1.15610747
#>  [4,] -0.94316692  0.69889684
#>  [5,]  0.21037607 -0.70891281
#>  [6,] -0.75611875  1.62289050
#>  [7,] -0.71515113  0.70916437
#>  [8,]  0.64108302  0.16285263
#>  [9,]  1.00716216 -0.45235596
#> [10,]  2.51424837 -0.48460600
#> [11,]  0.08691319  0.31282509
#> [12,]  0.63863058 -2.04986583
#> [13,]  0.32075453  0.75056431
#> [14,] -1.64639639  0.65621546
#> [15,]  0.06044466  0.44652192
#> [16,]  1.26779585 -1.17340837
#> [17,] -1.82140053 -0.02876119
#> [18,] -0.80678133 -1.00775856
#> [19,] -0.77285413 -0.40045805
#> [20,]  0.23624913  0.45551225